refactor: moved models to new module data.py
